+++ plugin/resources/assets/js/pages/popup/pages/PlagiarismPage.vue
@@ -2,42 +2,44 @@
     <div>
 
         <page-title title="Plagiarism">
-            <v-btn class="ma-2" tile outlined color="primary" @click="handleRunPlagiarismClicked">Run checksuite</v-btn>
         </page-title>
 
-        <plagiarism-results-section></plagiarism-results-section>
+        <plagiarism-matches-section @matchesFetched="passMatchesToOverview"></plagiarism-matches-section>
+
+        <plagiarism-overview-section :matches="matches"></plagiarism-overview-section>
+
+        <plagiarism-check-history-section></plagiarism-check-history-section>
 
     </div>
 </template>
 
 <script>
-    import {mapState} from 'vuex'
 
-    import {PageTitle} from '../partials'
-    import {PlagiarismResultsSection} from '../sections'
-    import {Plagiarism} from '../../../api'
+import {PageTitle} from '../partials'
+import {PlagiarismResultsSection, PlagiarismMatchesSection, PlagiarismOverviewSection} from '../sections'
+import PlagiarismCheckHistorySection from "../sections/PlagiarismCheckHistorySection";
 
-    export default {
-        name: 'plagiarism-page',
+export default {
+    name: 'plagiarism-page',
 
-        components: {PageTitle, PlagiarismResultsSection},
+    components: {
+        PlagiarismCheckHistorySection,
+        PlagiarismOverviewSection,
+        PlagiarismMatchesSection,
+        PageTitle,
+        PlagiarismResultsSection
+    },
 
-        computed: {
-            ...mapState([
-                'charon',
-            ]),
-        },
+    data() {
+        return {
+            matches: null
+        }
+    },
 
-        methods: {
-            handleRunPlagiarismClicked() {
-                Plagiarism.runPlagiarism(this.charon.id, response => {
-                    window.VueEvent.$emit(
-                        'show-notification',
-                        response.message,
-                        'success',
-                    )
-                })
-            },
+    methods: {
+        passMatchesToOverview(matchesFromMoss) {
+            this.matches = matchesFromMoss
         },
     }
+}
 </script>