+++ plugin/resources/assets/js/pages/popup/sections/index.js
@@ -13,6 +13,9 @@ import CharonSettingsSection from "./CharonSettingsSection";
 import DefenseRegistrationsSection from "./DefenseRegistrationsSection";
 import TesterTypeSection from "./TesterTypeSection";
 import TeacherSection from "./TeacherSection";
+import PlagiarismMatchesSection from "./PlagiarismMatchesSection";
+import PlagiarismOverviewSection from "./PlagiarismOverviewSection";
+import VisNetwork from "./VisNetwork";
 
 export {default as PlagiarismResultsSection} from './PlagiarismResultsSection'
 
@@ -20,5 +23,6 @@ export {
     CommentsSection, OutputSection, SubmissionOverviewSection,
     SubmissionsSection, LatestSubmissionsSection, LogSection, ActiveStudentsSection,
     SubmissionCountsSection, StudentsDistributionSection, ReportAllStudents, TeacherSection,
-    LabSection, CharonSettingsSection, DefenseRegistrationsSection, TesterTypeSection
+    LabSection, CharonSettingsSection, DefenseRegistrationsSection, TesterTypeSection, PlagiarismMatchesSection,
+    PlagiarismOverviewSection, VisNetwork
 }